Planchas ahora acepta decimales

+++ models.py
@@ -206,8 +206,8 @@ class KpiProductividad(Base):
     id = Column(Integer, primary_key=True, index=True)
     semana = Column(Integer, nullable=False)
     anio = Column(Integer, nullable=False)
-    meta_planchas = Column(Integer, nullable=False)
-    planchas_usadas = Column(Integer, nullable=False)
+    meta_planchas = Column(Numeric(10,2), nullable=False)
+    planchas_usadas = Column(Numeric(10,2), nullable=False)
 
 class KpiVentas(Base):
     __tablename__ = "kpi_ventas"
